+++ RELEASES.md
@@ -44,6 +44,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template: template variable syntax checker
 * prov-template-compiler: introduced fully qualified names for templates
 * prov-template-compiler: LocalEnactor for composites
+* prov-template-compiler: support for template library paths
 
 `2.2.1`
 * service-metrics: a service to gather metrics on provenance, using MetricsCalculator.java